backend/requirements.txt
# Core FastAPI dependencies
fastapi
uvicorn[standard]
pydantic
python-dotenv

# HTTP client
httpx

# Additional utilities
python-multipart
pydantic-settings